Fixed bug where Constituent ID's in a parse were not unique.
There was a bug in s2chelper() that caused Constituent ID's to be
repeated.  s2chelper() calls itself recursively.  The 'idCounter'
parameter was being incremented and then passed by *value* instead of
passed by *reference*.  The source code had this comment, explainging
how the counter was being passed by reference:

  /**
   * i'm using int[] as a java hack for int* (pass by reference rather than value).
   */

but the current version of the code was not using this Java hack.

It looks like the bug was introduced as part of the transition from
protobufs to Thrift sometime in November 2013.

I modified s2chelper() to match the behavior of the (working) protobuf
version of the function.