Required fields are now set when AGIGA doc has no Entities or Mentions
When the AGIGA input document had no entities or mentions, the
Concrete document created by concrete-agiga would have:
 - an EntityMentionSet with no mentionSet field
 - an EntitySet with no entityList field

Per the Concrete schema, it's okay for a Communication not to have an
EntityMentionSet, or for EntityMentionSet.mentionSet to be an empty
list - but not for EntityMentionSet.mentionSet to be undefined.

The mentionSet and entityList fields are now empty lists when the
AGIGA source document has no entities or mentions.